# internal/transport/handler/docs/swagger.yaml
basePath: /
definitions:
  model.Delivery:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    type: object
  model.Item:
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        type: integer
      price:
        type: integer
      rid:
        type: string
      sale:
        type: integer
      size:
        type: string
      status:
        type: integer
      total_price:
        type: integer
      track_number:
        type: string
    type: object
  model.Order:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/model.Delivery'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        type: string
      payment:
        $ref: '#/definitions/model.Payment'
      shardkey:
        type: string
      sm_id:
        type: integer
      track_number:
        type: string
    type: object
  model.Payment:
    properties:
      amount:
        type: integer
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        type: integer
      delivery_cost:
        type: integer
      goods_total:
        type: integer
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Service for processing, storing and displaying order data
  title: Order service API
  version: "1.0"
paths:
  /:
    get:
      description: Returns an HTML page with an order search form.
      produces:
      - text/html
      responses:
        "200":
          description: HTML content
          schema:
            type: string
      summary: Display web page
      tags:
      - page
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Returns an order by its unique identifier (UID).
      parameters:
      - description: Order UID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get order data by ID
      tags:
      - orders
swagger: "2.0"